pyproject.toml
[build-system]
requires = ["pdm"]

[project]
name = ""
version = "0.0.1"
description = ""
authors = [
    {name = "Marc Steiner", email = "msteiner@ini.uzh.ch"},
]
dependencies = [
    "pandas>=1.5",
    "numpy>=1.22",
    "plotly>=5.11.0",
    "torch>=1.13.1",
    "torchvision>=0.14.1",
    "pytorch-lightning==1.9.1",
    "torchmetrics==0.11.0",
    "hydra-core==1.3.1",
    "hydra-colorlog==1.2.0",
    "hydra-optuna-sweeper==1.2.0",
    "hydra-joblib-launcher>=1.2.0",
    "kaleido==0.2.1",
    "pip>=22.3.1",
    "scikit-learn>=1.2.0",
    "mypy>=0.991",
    "pandas-stubs>=1.2.0.62",
    "torchinfo>=1.7.1",
    "torch-optimizer>=0.3.0",
    "numba==0.56",
    "pyrootutils>=1.0.4",
    "pre-commit>=3.1.0",
    "wandb>=0.13.10",
    "rich>=13.3.1",
    "pytest>=7.2.1",
    "tensorboard>=2.12.0",
]
requires-python = ">=3.8"
license = {text = "MIT"}

[tool.pdm]

[tool.pytest.ini_options]
addopts = [
  "--color=yes",
  "--durations=0",
  "--strict-markers",
  "--doctest-modules",
]
filterwarnings = [
  "ignore::DeprecationWarning",
  "ignore::UserWarning",
]
log_cli = "True"
markers = [
  "slow: slow tests",
]
minversion = "6.0"
testpaths = "tests/"

[tool.coverage.report]
exclude_lines = [
    "pragma: nocover",
    "raise NotImplementedError",
    "raise NotImplementedError()",
    "if __name__ == .__main__.:",
]
